refactor: make initialization of i18next extendable (#16662)
Previously, the initialization of the `i18next` instance happened in a private and non-extendable function `i18nextInit()`. This function was hard to customize, because it was run in an `APP_INITIALIZER` provided in a deeply-nested `I18nModule` (nested inside `BaseStorefrontModule`). When a customer wanted to customize the logic of initializing `i18next` (e.g. use a different i18next initialization options or use a different strategy for lazy-loading JSON translations - for example [load from filesystem instead from http](https://github.com/SAP/spartacus/issues/13460)), then they had to "un-import" the whole tree of `BaseStorefrontModule` and dismanle it, then opt out from our default deeply-nested `I18nModule`, and then re-assemble the remaining modules tree.

Now the logic is extendable, and it's very easy to customize. The initialization of i18next and defining strategies for loading translations from backend are split into a few injectable classes with different responsibilities:

![image](https://user-images.githubusercontent.com/4001059/213250056-6b810bcf-0585-440d-9ac8-05ac70d0ce10.png)

- `I18nextInitializer.initialize()` - main function to initialize `i18next`; called in `APP_INITIALIZER`
- `I18nextBackendService` - abstract class injection token, configures some i18next backend plugin. It's a dependency of `I18nextInitializer`
- `I18nextHttpBackendService` - A specific implementation of `I18nextBackendService` token. It configures specific i18next backend plugin (for loading translations via HTTP); note: there can be potentially [other possible i18next backend plugins](https://www.i18next.com/overview/plugins-and-utils#backends) used (e.g.  [in the future we might want to load translations from a filesystem in SSR](https://github.com/SAP/spartacus/issues/13460) by using `i18next-fs-backend` plugin only in SSR). 
- `I18NEXT_HTTP_BACKEND_CLIENT` - injectable function used for loading translations via http; under the hood it uses Angular `HttpClient` (to work properly in SSR). It's a dependency of specific `I18nextHttpBackendService`.

Note: Here's the followup PR, that uses `i18next-fs-backend` in SSR: https://github.com/SAP/spartacus/pull/16791

closes [CXSPA-2061](https://jira.tools.sap/browse/CXSPA-2061)
closes #11675

---
**QA steps**
This is just a refactor. So everything should work as before (in the `develop` branch).

**Environments to test:**
- browser (`yarn start`)
- ssr (`yarn dev:ssr` + open a site in the browser but [with Javascript disabled](https://developer.chrome.com/docs/devtools/javascript/disable/))

**i18next configurations to test:**
- static translations (used by default)
- external translations. Here are prerequisites:
  - generate JSON translations and copy them to the folder `/projects/storefrontapp/src/assets`:
  ```bash
  yarn build:assets
  cp -r dist/assets/i18n-assets projects/storefrontapp/src/assets
  ```
  - in app.module configure Spartacus to load JSON translations instead of static resources. In app.module please change (please make sure to delete the line with `resources`!):
  ```diff
        i18n: {
  -       resources: translations,
  +       backend: {
  +         loadPath: 'assets/i18n-assets/{{lng}}/{{ns}}.json',
  +       },
  ```