feat: improve test stability (CXSPA-2619) (#16933)
This should reduce test flakiness.
Pricing requests are fired when the get/update config returns. A cy.wait() will wait for the last fired request, matching the given route, that has not yet been waited for. In case of pricing we do not wait for each pricing request in all test cases.
So the following might happen. We typically start to wait for the pricing request, after the update request returns. Because the pricing is fired when the update returns, it might be that pricing request actually fires a little bit after the cypress test started to wait for it. In this case it will most likely find an 'old' pricing request that has not been waited for and the wait immediately returns as it only waits for the 'old' pricing request that did already return.
By just delaying (100ms) the wait for pricing requests, we ensure that the pricing requests was fired, so that the wait will wait for the correct request.
For one test class 'product-configurator-tabbing.core-e2e-spec.ts' this PR ensures that we wait for all pricing requests, hence in this case the delay of 100ns could be removed. However this is not implemented for all product config test classes, yet. hence the delay is still required.