chore: fix regression in command service error handling (#17452)
CXSPA-3535

The regression is that errors thrown by the commandFactory would not be caught and result in the processing stream closing.  Future command requests would not be processed.  The update wraps the commandFactory function to catch any thrown errors and pass them to the subscriber.  It additionally prevents the Error-previous strategy from overwriting any notified errors with the "Canceled by next command" error, an edge-case which would occur for synchronous commands only.