fix: force creation of a new CPQ session if a new token is issued (CXSPA-3496) (#17488)
Once a CPQ session has been created with the business context from the token, this context can not be changed anymore. So if a new token is requested for a different customer we also new to ensure that a new CPQ session is created. Reusing any existing session may not work.