fix: prevent `CmsService.getComponentData()` invoked with an empty string `uid` from unnecesarily calling Ngrx and Backend (#19229)
Previously:
When the `CmsService.getComponentData()` was invoked with an empty string `uid` argument, then it called Backend (and also Ngrx along the way), which resulted in a Failure Ngrx action. And it emitted `null` value. 

See `componentIds=` in the backend URL in the screenshot below:
<img width="1217" alt="image" src="https://github.com/user-attachments/assets/a6f05dde-5c72-4fb8-81d6-784ef6703afb">


Observations:
- It doesn't make sense to make a call to Backend with and empty `uid` - we want to prevent it.
- The Failure Ngrx action happening during SSR, made the SSR of PDP fail (return 500) - which breaks SEO - so we want to prevent it.

Fix:
Now we validate the `uid` argument of `CmsService.getComponentData()` and emit `null` immediately if the `uid` is empty. In such case, we don't call the backend anymore (would caused getting empty response from Backend and a failure from Ngrx).

fixes https://jira.tools.sap/browse/CXSPA-8389