fix: rendering of focus indicator not overlapping with other element CXSPA-7955) (#19246)
The hamburger menu button just used padding for spacing, hence the focus indicator could overlap with other elements and looked a bit odd. Changed the spacing to be distributed between margin and padding, hence the focus indicator will render nicely with 5px white space around the control and not overlap.