chore(schematics/testing.ts): restore previous `@spartacus:registry` instead of hardcoded NPM registry on the process exit (#19378)
Currently, on the process exit, the value of `@spartacus:registry` is restored to the outdated URL of public NPM registry `https://registry.npmjs.org/`. It was useful only until we stopped releasing to public NPM many months ago. Now it's confusing and a bit annoying. It's because you'd rather like it to be restored to a previous real registry url, e.g. RBSC.

To fix it, now we remember the used `@spartacus:registry` before running the script `schematics/testing.ts`. Then the scrip overwrites `@spartacus:registry` to verdaccio URL (as it does usually). And then on the process exit we restore the **remembered URL** original URL. ... which should improve DX while re-running this script multiple times.

closes https://jira.tools.sap/browse/CXSPA-8636