chore: bump RxJS from 7.8.1 to 7.8.2 only in our package-lock.json (no peerDeps update) (#20309)
This PR bumps the globally **installed** version of rxjs in our repo from 7.8.1 to 7.8.2.
However we keep the minimal requirements for customers on the level `^7.8.0`.

Unfortunately the installed version of `@angular-devkit/schematics/testing` has an internalized copy of rxjs in `node_modules/@angular-devkit/schematics/node_modules/rxjs` which still has version 7.8.1.

The typings between rxjs@7.8.2 and rxjs@7.8.1 are a bit incompatible, which caused typing issues in unit tests of schematics, when calling the function `firstValueFrom` imported from the globally installed `rxjs@7.8.2` , called with an argument  `schematicRunner` imported from `@angular-devkit/schematics/testing` (which is based on the internalized version of rxjs `node_modules/@angular-devkit/schematics/node_modules/rxjs@7.8.1`).

I've tried to bump just `@angular-devkit/schematics` to a higher version (so it starts using a higer internalized version of rxjs 7.8.2, but not 7.8.1). I also tried upgrading it all other Angular libs at once (to the latest patch 19.0.7 from the current 19.0.4), but had problem to proceed with it without peerDependencies errors. 
Later on I've checked in aside "empty test npm repo" that @angular-devkit/schematics 19.0.7 (highest patch on 19.0.x version line) uses rxjs@7.8.1 - so even bumping it in our repo to 19.0.7 wouldn't help

So I decided to use the following workaround:
In those unit tests which used the rxjs function `firstValueFrom`, we're importing the function `firstValueFrom` not from the globally installed `rxjs@7.8.2`, but from th internalized version of rxjs inside `@angular-devkit/schematics/node_modules/rxjs@7.8.1` to be type-compatible with the `schematicRunner` from `@angular-devkit/schematics/testing`.

Please note the workaround is needed only in unit tests, not in the source code of our libs. The public API that we ship to our customers didn't change.

This workaround can be removed in the future, when we'll bump `@angular-devkit/schematics` (which will happen at latest during our yearly February Framework Update Release, when we update all Angular packages).

fixes https://jira.tools.sap/browse/CXSPA-10172